Minor code review update

+++ d3d12game_win32_dr/pch.h
@@ -62,7 +62,7 @@ namespace DX
         virtual const char* what() const override
         {
             static char s_str[64] = { 0 };
-            sprintf_s(s_str, "Failure with HRESULT of %08X", result);
+            sprintf_s(s_str, "Failure with HRESULT of %08X", static_cast<unsigned int>(result));
             return s_str;
         }
